+++ src/browserist/browser/check_if/is_clickable.py
@@ -7,7 +7,7 @@ from ...model.browser.base.driver import BrowserDriver
 from ...model.type.xpath import XPath
 
 
-def check_if_is_clickable(browser_driver: BrowserDriver, xpath: str, timeout: int) -> bool:
+def check_if_is_clickable(browser_driver: BrowserDriver, xpath: str, timeout: float) -> bool:
     xpath = XPath(xpath)
     driver = browser_driver.get_webdriver()
     try:
