+++ src/browserist/browser/combo/log_in.py
@@ -12,11 +12,11 @@ from ..wait.for_element import wait_for_element
 from ..wait.until.url.contains import wait_until_url_contains
 
 
-def combo_log_in(driver_method: DriverMethods, login_credentials: LoginCredentials, login_form: LoginForm1Step | LoginForm2Steps, timeout: int) -> None:
+def combo_log_in(driver_method: DriverMethods, login_credentials: LoginCredentials, login_form: LoginForm1Step | LoginForm2Steps, timeout: float) -> None:
     timeout_should_continue: TimeoutShouldContinueCallable = driver_method._timeout_should_continue
     browser_driver: BrowserDriver = driver_method._browser_driver
 
-    def login_form_1_step(browser_driver: BrowserDriver, timeout_should_continue: TimeoutShouldContinueCallable, login_form: LoginForm1Step, timeout: int) -> None:
+    def login_form_1_step(browser_driver: BrowserDriver, timeout_should_continue: TimeoutShouldContinueCallable, login_form: LoginForm1Step, timeout: float) -> None:
         if timeout_should_continue():
             input_value(browser_driver, login_form.username_input_xpath, login_credentials.username, timeout)
         if timeout_should_continue():
@@ -24,7 +24,7 @@ def combo_log_in(driver_method: DriverMethods, login_credentials: LoginCredentia
         if timeout_should_continue():
             click_button(browser_driver, login_form.submit_button_xpath, timeout)
 
-    def login_form_2_steps(browser_driver: BrowserDriver, timeout_should_continue: TimeoutShouldContinueCallable, login_form: LoginForm2Steps, timeout: int) -> None:
+    def login_form_2_steps(browser_driver: BrowserDriver, timeout_should_continue: TimeoutShouldContinueCallable, login_form: LoginForm2Steps, timeout: float) -> None:
         if timeout_should_continue():
             input_value(browser_driver, login_form.username_input_xpath, login_credentials.username, timeout)
         if timeout_should_continue():
